use assetHooks, clean up
